refactor: `VItemRenderer`
